requirements.txt
nose==1.3.7
numpy==1.9.2
scipy==0.16.0
matplotlib==1.4.3
wget==3.2
pyyaml==3.11